Extract CustomActions mixin
